refactor: formatting

+++ server/src/api/inventory/labs/delete.ts
@@ -7,9 +7,13 @@ import { Router } from "express";
 
 const router = Router();
 
-router.delete('/:id', checkAccess(), asyncHandler(async (req, res) => {
-    await db.delete(laboratories).where(eq(laboratories.id, req.params.id));
-    res.status(200).json({ message: 'Laboratory deleted successfully' });
-}));
+router.delete(
+    "/:id",
+    checkAccess(),
+    asyncHandler(async (req, res) => {
+        await db.delete(laboratories).where(eq(laboratories.id, req.params.id));
+        res.status(200).json({ message: "Laboratory deleted successfully" });
+    })
+);
 
 export default router;
